feat(sdk)!: type the size APIs size_t, and give cDescription a size_t capacity (#294)
`-count`, `-length`, an index, a capacity and the `count:` of a literal
builder are sizes, and were spelled `unsigned int`. They are `size_t` now,
which is what the concept is called in C and what Zephyr's own APIs use.
`OZHeap`'s `-usedBytes` already did, so the SDK was inconsistent with
itself.

`-cDescription:maxLength:` becomes `int (char *, size_t)` -- `snprintf`'s
own shape, which is the function it imitates. The capacity cannot be
negative, so it is a `size_t`; the result stays `int`, and the two
alternatives are worth recording because both were tried:

  - `ssize_t` is the natural spelling and is unusable. It lives in
    `<sys/types.h>`, and the Clang AST dump that decides ivar ownership
    runs `--target=x86_64-unknown-linux-gnu` with no Linux sysroot, so only
    *compiler-provided* headers resolve there. `<stddef.h>` and
    `<stdint.h>` are Clang's own; `<sys/types.h>` is libc's. Including it
    broke all 74 behaviour cases at the dump step.
  - `ptrdiff_t` compiles on both paths but means "difference between two
    pointers", which is not what the method returns. An Objective-C
    interface saying so reads wrong, and it would have said so in eight
    places including two samples people read as examples.

The accumulating implementations needed reading, not just retyping: they
pass `maxLength:maxLen - pos` recursively. That subtraction is safe
unsigned only because every one of them is guarded by `pos < maxLen`, and
`pos` can reach `maxLen` but never pass it -- each multi-byte write is
guarded by `pos + n < maxLen`. Those guards are load-bearing now rather
than incidental, and the code says so.

The accumulation is also guarded, which it was not before:

    int written = [elem cDescription:buf + pos maxLength:maxLen - pos];
    if (written > 0) {
            pos += (size_t)written;
    }

A negative result used to be added straight into a signed `pos` and then
used to index `buf`.

`_oz_q31_to_str`'s `if (maxLen <= 0)` became `== 0`: an unsigned capacity
has no negative to reject.

Callers followed: every class in the samples, the behaviour corpus, the
adapted suites and the Rust tests that declares a count, a length or an
override of `cDescription:`, the `%u` format specifiers that are now `%zu`,
and one marker string in `tests/common/mod.rs` that slices the assembled
SDK header by its literal type spelling.

`OZQ31`'s `-unsignedIntValue` and `+fixedWithUnsignedInt:` are untouched:
they extract a *value*, not a size.

Marked `!`: these signatures changed.

+++ include/oz_sdk/Foundation/OZArray.h
@@ -13,21 +13,21 @@ struct NSFastEnumerationState;
 
 @interface OZArray<__covariant ObjectType> : OZObject <IteratorProtocol> {
 	__unsafe_unretained id *_items;
-	unsigned int _count;
+	size_t _count;
 	uint16_t _iterIdx;
 }
 
 @property (readonly) uint16_t iterIdx;
 
-+ (id)arrayWithObjects:(const id *)objects count:(unsigned int)count;
-- (unsigned int)count;
-- (id)objectAtIndex:(unsigned int)index;
-- (id)objectAtIndexedSubscript:(unsigned int)index;
-- (void)enumerateObjectsUsingBlock:(void (^)(id obj, unsigned int idx, BOOL *stop))block;
++ (id)arrayWithObjects:(const id *)objects count:(size_t)count;
+- (size_t)count;
+- (id)objectAtIndex:(size_t)index;
+- (id)objectAtIndexedSubscript:(size_t)index;
+- (void)enumerateObjectsUsingBlock:(void (^)(id obj, size_t idx, BOOL *stop))block;
 - (unsigned long)countByEnumeratingWithState:(struct NSFastEnumerationState *)state
 				     objects:(__unsafe_unretained id *)stackbuf
 				       count:(unsigned long)len;
-- (int)cDescription:(char *)buf maxLength:(int)maxLen;
+- (int)cDescription:(char *)buf maxLength:(size_t)maxLen;
 - (instancetype)iter;
 - (id)next;
 @end
